Fixed retrieval of AWS regions.
We recently found an issue in CloudX team, we had hundreds of AMIs
running and not cleaned up properly by cloud-cleaner.

After debugging, we found out that the regions affected were the ones
filtered out by the current implementation of this script, which gathers
only the regions enabled by default.

By using the OptInStatus != "not-opted-in" filter, we solve this issue
and we gather all the regions actually enabled in the Account,
despite they were enabled by default or opted in later on.