aws: drop s3 cleaning
S3 supports lifecycle policies which can do the same thing as
our script: delete old objects based on their age (with support for
filtering based on tags).

Let's use these instead so we don't have to do it using this script.
This allows us to just delete all code related to S3 from this repo.
Since this was the last piece using awscli, we can also remove the
awscli initialization.